feat: reorganización de los modulos de aplicación en una sola carpeta contenedora. udistrital/sisifo_documentacion#741
