+++ src/application/auditor/auditor.service.spec.ts
