+++ src/application/auditor/auditor.service.ts
