+++ src/application/auditoria/auditoria.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AuditoriaController } from './auditoria.controller';
 import { AuditoriaService } from './auditoria.service';
-import { AuditorModule } from '../../auditor/auditor.module';
+import { AuditorModule } from '../auditor/auditor.module';
 import { DominiosModule } from 'src/shared/utils/dominios/dominios.module';
 import { ServicesModule } from 'src/shared/services/services.module';
